% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.name.function.R
\name{check.name}
\alias{check.name}
\title{Check first column of data frame for names}
\usage{
check.name(df)
}
\arguments{
\item{df}{A data frame}
}
\value{
A data frame (with first column removed if values were not integers or numerics)
}
\description{
If values in first column of data frame are not integers or numerics, the first column will be erased.
}
