% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/G2P.function.R
\name{G2P}
\alias{G2P}
\title{Simulate a phenotype}
\usage{
G2P(X = NULL, h2 = 0.5, alpha = 1, NQTN = 10, distribution = "normal", a2 = 0)
}
\arguments{
\item{X}{A genotype matrix (rows = taxa, columns = markers).}

\item{h2}{The heritability of the simulated trait.}

\item{alpha}{The p parameter in an approximated geometric distribution}

\item{NQTN}{The number markers contributing to the phenotype (if known)}

\item{a2}{The proportion of non-additive interactive variance}
}
\value{
A list of the marker effects (1*NQTN), the phenotype (n*1), the combined additive effects (n*1), the residual effects (n*1), the additive marker indices (1*m), and the interaction marker positions (1*nint)
}
\description{
Simulate a phenotypic data from genotypic data.
}
