% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GLM.function.R
\name{GLM.func}
\alias{GLM.func}
\title{GWAS by GLM}
\usage{
GLM.func(geno = NULL, pheno = NULL, covariates = NULL, PCs = 1, thresh = 0.2)
}
\arguments{
\item{geno}{A matrix of genotype data with dimensions n x m}

\item{pheno}{A matrix of phenotype data with dimensions n x 1}

\item{covariates}{A matrix of covariate data with dimensions n x t}

\item{PCs}{A matrix of principle components with dimensions n x ? (variable number of PCs)}

\item{thresh}{A numeric (0-1) indicating the correlation value of any PC-covariate pair at which the PC will be exluded from model}
}
\value{
A matrix of p values with dimensions 1 x m
}
\description{
Perform GWAS using a general linear model method incorporating principle components and additional covariates if present.
}
