% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GWAS.by.cor.function.R
\name{GWASbyCor}
\alias{GWASbyCor}
\title{Perform GWAS by Correlation}
\usage{
GWASbyCor(X, y)
}
\arguments{
\item{X}{A genotype matrix with dimensions n x m (rows = taxa, columns = markers).}

\item{y}{A phenotype vector of length n}
}
\value{
A vector of p-values of length m
}
\description{
Perform GWAS by Correlation
}
