% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GWASP.R
\docType{package}
\name{GWASP}
\alias{GWASP}
\title{GWASP: A package for performing GWAS by GLM using PCs as covariates.}
\description{
The GWASP package performs GWAS by GLM using PCs as covariates to account for population structure.
}
