% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.name.function.R
\name{check.name}
\alias{check.name}
\title{Check first column of matrix for names}
\usage{
check.name(df)
}
\arguments{
\item{df}{A data frame}
}
\value{
A data frame with row names from the first column if the first column was not numeric or integer
}
\description{
If values in first column are not integers or numerics, the first column will be changed to a row name.
}
