% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GLM.function.R
\name{GLM.func}
\alias{GLM.func}
\title{GWAS by GLM}
\usage{
GLM.func(X = NULL, y = NULL, C = NULL, PCs = 1, r = 0.2)
}
\arguments{
\item{X}{A matrix of genotype data with dimensions n x m}

\item{y}{A matrix of phenotype data with dimensions n x 1}

\item{C}{A matrix of covariate data with dimensions n x t}

\item{PCs}{A matrix of principle components with dimensions n x ? (variable number of PCs)}

\item{r}{A numeric (0-1) indicating the correlation value of any PC-covariate pair at which the PC will be exluded from model}
}
\value{
A matrix of p values with dimensions 1 x m
}
\description{
Perform GWAS using a general linear model method incorporating principle components and additional covariates if present.
}
